add makeModel
